Workarounds to deal with apparent clang limitation that clang doesn't handle passing compile time const expressions through a static inline function argument.
This has the unfortunate side effect of either noticeable degrading
performance of HPC simd code or making HPC simd code much harder to
implement / maintain / port to non-x86.

This includes sig-verify.  This practically means, as far as Firedancer
is concerned, clang is currently mostly useful for code hygenine
purposes and is not a particularly practical alternative to gcc
otherwise.

Change-Id: I42ccd59f26d5e82bb9c88a66f9d58743bab62f4c
Reviewed-on: https://git.firedancer.io/c/firedancer/+/247
Reviewed-by: Kevin Bowers <kbowers@jumptrading.com>