Another binary Merkle tree belt sanding pass
This added new functionality, further simplified the usage and
implementation and has some additional performance improvements.

- Users do not need to specify the number of leaves up front.  This
  allowed simplified constructor APIs, elimination of all edge cases
  under the hood, and allowed for commit state to be just simplify
  declared at compile time on the stack and/or data segment.
  Implementation still supports arbitrary sized trees (at least for
  trees that are computable on human-timescales and easily extensible if
  wanting to go further).

- Related to the above, added an API for getting the the number of
  leaves already appended to a calculation.

- Tweaked data structure layouts to eliminate additional unnecessary
  copies and make it easier for the compiler to hardware accelerate
  and/or add more advanced hardware accelerations in the future.

An interesting side effect is that, at this point, there is practically
no difference under the hood between the different variants.  So, if
wanting to go further, we could in detemplatized the code and simplify
the naming too (not done currently to minimize impact on existing code).