Minor naming cleanup
Renamed fd_funk_txn_id_t to fd_funk_txn_xid_t (and related tokens) to
disambiguate from other uses of the name id and be more lexically
consistent with fd_funk_rec_key_t.  Touches a lot of files in here in
tiny ways but better to do it now than later (when it is too late).