pod: fix prefix as 'firedancer'
This is there to namespace firedancer configuration from external
programs or users who want to add their own things to the pod. It is
not to namespace Firedancer instances from each other, which will
already run in different workspaces. Fixing the prefix removes one
piece of information needed to lookup configuration if all you have
is the workspace.