fdctl: diagnose if nic does not support channels
Consumer grade NICs often don't support setting the number of channels
but this is a requirement if we are running with more than one QUIC
tile. We add a helpful diagnostic message in this case as it might
be quite common when running on development boxes, and also suggest
a potential fix for development (running with one QUIC tile).