fdctl: create topology and rewire tile init
The lack of a consistent view of the tile topology was creating some
headaches.  We couldn't know in advance the total memory being used, so
workspace sizing was manual and error prone.  As well, any time the code
needed to ask questions like "how many tiles are there?" it was
difficult to answer.

A new topology type is introduced which represents the overall layout
of Firedancer, it consists of workspaces, tiles and links.  Tiles and
links go inside workspaces, and links go between tiles.  All of our
queries like "how many tiles are there?" or "who consumes this data?"
are now answered by the topology.