fctl: give each tile unique adress space layout
The boot path is rewired to use pipes and SIGHUP for determining when a
process is dead.  This is needed to fix some rare cases where orphaned
processes would be leftover.

Then the boot sequence is switched so that tiles are started via. execve
rather than clone.  This boot sequence is now:

 1. The main process starts.
    - It creates a pipe, main_pipe, with the read side CLOEXEC.
    - It spawns a child process, pidns with clone(CLONE_NEWPID).  This
      process exists solely to be the root of a PID namespace. The read
      side of main_pipe is closed in pidns and it holds onto the write
      side.
    - The main process closes the write side of main_pipe.

Now main and pidns can tell if the other one died by polling for HUP on
the pipe fd.  Main doesn't need to do this since it can call waitpid()
on the pidns pid, but the pidns does need to poll, in-case main is
killed with a SIGKILL.

  2. The pidns process starts.
    - It spawns each tile one by one, by calling fork() and then
      execve(). Before spawning the tile it creates a pipe, tile_pipe,
      with the read side CLOEXEC.  It passes the write side to the child
      (tile) when it is execve'd.
    - Because memory maps are lost on an execve, but it needs to share
      the configuration with child tiles, it creates a memfd with
      memfd_create, writes the config there, passes the fd through the
      execve call, and has the tile map the memfd in and read it.
    - A similar memfd sequence is done with the lock object for
      sequencing writes the log file.
    - The pidns process closes the write side of tile_pipe.

The pidns process can now call poll( main_read, tile_pipe0_read,
tile_pipe1_read, ... ) to wait for any the parent or child processes to
die.  After that happens it will exit itself, and because it's the init
of a PID namespace, all of its children will then be terminated as well
by the kernel.

   3. Each tile process starts.
    - It immediately calls clone() to spawn a new process with a
      stack backed by huge pages.  The tile_pipe write side is
      passed onto the cloned child.
    - It returns so that the non-hugepage process exits.
    - The pidns process reaps the non-hugepage process.

Now we finally have all our tile processes ready, running on huge pages,
with a unique address space layout because they were restarted via.
execve.  The pidns and main processes are linked by a pipe, and the
pidns and child tiles are linked by a pipe as well, meaning that if any
process in the tree exits, everything else will get taken down cleanly.