metrics: set up infrastructure for collection
Primarily metrics will be used for aggregating high frequency events
which can't be reported on an individual basis.  Lower frequency events
like consensus related information (we finished a leader slot) will
not want to aggregate, and use separate infrastructure.

The metric types are designed to align with prometheus and otel and
a future change will expose them via. a local prometheus endpoint.