util: don't zero entire wksp or dcache
With a large workspace or dcache, zeroing the entire space can take a
long time.  This isn't strictly necessary, since the wksp and dcache
implementations do not assume the data region is zeroed.

The underlying shared memory must be zeroed by the kernel when it is
first grabbed, so the only way this is non-zero is if we are reusing
space in the shared memory that we created previously.  It is OK to
expose, but not use, the previous data in this case.